added fullname, username, and profile pic visiblity.
added swift UserModel to retrieve user data, made edits to PostView (ie: Auto keyboard and changed button images, added data to be sent to Firebase on Google sign in and Email sign up. Note: We need to change how often data is updated on Google Sign in. It should only be once when a username and fullname, profile image, is sent to firebase.